+++ src/model/Model_Translink.cpp
@@ -50,7 +50,7 @@ Model_Translink& Model_Translink::instance()
     return Singleton<Model_Translink>::instance();
 }
 
-Model_Translink::CHECKING_TYPE Model_Translink::type_checking(const int tt)
+Model_Translink::CHECKING_TYPE Model_Translink::type_checking(const int64 tt)
 {
     if (tt == AS_TRANSFER)
     {
@@ -62,25 +62,25 @@ Model_Translink::CHECKING_TYPE Model_Translink::type_checking(const int tt)
     }
 }
 
-Model_Translink::Data* Model_Translink::SetAssetTranslink(const int asset_id
-    , const int checking_id
+Model_Translink::Data* Model_Translink::SetAssetTranslink(const int64 asset_id
+    , const int64 checking_id
     , const CHECKING_TYPE checking_type)
 {
     return SetTranslink(checking_id, checking_type
         , Model_Attachment::reftype_desc(Model_Attachment::ASSET), asset_id);
 }
 
-Model_Translink::Data* Model_Translink::SetStockTranslink(const int stock_id
-    , const int checking_id
+Model_Translink::Data* Model_Translink::SetStockTranslink(const int64 stock_id
+    , const int64 checking_id
     , const CHECKING_TYPE checking_type)
 {
     return SetTranslink(checking_id, checking_type
         , Model_Attachment::reftype_desc(Model_Attachment::STOCK), stock_id);
 }
 
-Model_Translink::Data* Model_Translink::SetTranslink(const int checking_id
+Model_Translink::Data* Model_Translink::SetTranslink(const int64 checking_id
     , const CHECKING_TYPE checking_type
-    , const wxString& link_type, const int link_record_id)
+    , const wxString& link_type, const int64 link_record_id)
 {
     Model_Translink::Data* translink = Model_Translink::instance().create();
     translink->CHECKINGACCOUNTID = checking_id;
@@ -100,7 +100,7 @@ Model_Translink::Data* Model_Translink::SetTranslink(const int checking_id
 }
 
 Model_Translink::Data_Set Model_Translink::TranslinkList(Model_Attachment::REFTYPE link_table
-    , const int link_entry_id)
+    , const int64 link_entry_id)
 {
     Model_Translink::Data_Set translink_list = Model_Translink::instance().find(
         Model_Translink::LINKTYPE(Model_Attachment::reftype_desc(link_table))
@@ -109,7 +109,7 @@ Model_Translink::Data_Set Model_Translink::TranslinkList(Model_Attachment::REFTY
     return translink_list;
 }
 
-bool Model_Translink::HasShares(const int stock_id)
+bool Model_Translink::HasShares(const int64 stock_id)
 {
     if (TranslinkList(Model_Attachment::STOCK, stock_id).empty())
     {
@@ -119,7 +119,7 @@ bool Model_Translink::HasShares(const int stock_id)
     return true;
 }
 
-Model_Translink::Data Model_Translink::TranslinkRecord(const int checking_id)
+Model_Translink::Data Model_Translink::TranslinkRecord(const int64 checking_id)
 {
     auto i = Model_Translink::CHECKINGACCOUNTID(checking_id);
     Model_Translink::Data_Set translink_list = Model_Translink::instance().find(i);
@@ -132,7 +132,7 @@ Model_Translink::Data Model_Translink::TranslinkRecord(const int checking_id)
     }
 }
 
-void Model_Translink::RemoveTransLinkRecords(Model_Attachment::REFTYPE table_type, const int entry_id)
+void Model_Translink::RemoveTransLinkRecords(Model_Attachment::REFTYPE table_type, const int64 entry_id)
 {
     for (const auto& translink : TranslinkList(table_type, entry_id))
     {
@@ -140,7 +140,7 @@ void Model_Translink::RemoveTransLinkRecords(Model_Attachment::REFTYPE table_typ
     }
 }
 
-void Model_Translink::RemoveTranslinkEntry(const int checking_account_id)
+void Model_Translink::RemoveTranslinkEntry(const int64 checking_account_id)
 {
     Data translink = TranslinkRecord(checking_account_id);
     Model_Shareinfo::RemoveShareEntry(translink.CHECKINGACCOUNTID);
@@ -245,7 +245,7 @@ void Model_Translink::UpdateAssetValue(Model_Asset::Data* asset_entry)
     }
 }
 
-bool Model_Translink::ShareAccountId(int& stock_entry_id)
+bool Model_Translink::ShareAccountId(int64& stock_entry_id)
 {
     Model_Translink::Data_Set stock_translink_list = TranslinkList(Model_Attachment::STOCK, stock_entry_id);
